Fixed computation with null eigenvalues. Fixed print of variables on windows.
git-svn-id: https://svn.vital-it.ch/svn/hp2c/trunk/Codeml_Variants/Fastcodeml@3351 95c5a404-1f88-0410-a6b0-c3f062b6f34a

+++ CodeMLoptimizer.cpp
@@ -63,7 +63,7 @@ static const int MAX_ITERATIONS=10000;
 
 double Ming2::minimizeFunction(unsigned int aFgBranch, std::vector<double>& aVars)
 {
-	unsigned int np = aVars.size();
+	int np = (int)aVars.size();
 	std::vector<double> space(np*(np*2+9+2));
 	std::vector<int> ispace(2*np);
 	const double e = 1e-8;
